requirements.txt
absl-py==2.3.1
ale-py==0.11.2
alembic==1.16.4
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
Farama-Notifications==0.0.4
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.4
grpcio==1.74.0
gym-notices==0.1.0
gymnasium==1.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
optuna==4.4.0
packaging==25.0
pandas==2.3.1
pettingzoo==1.24.3
pillow==11.3.0
protobuf==6.31.1
psutil==7.0.0
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
sb3_contrib==2.7.0
setuptools==80.9.0
Shimmy==2.0.0
six==1.17.0
SQLAlchemy==2.0.42
stable_baselines3==2.7.0
-e git+https://github.com/LucasAlegre/sumo-rl@41e4c63eb4652c27ec18bc553048bb251f01cddf#egg=sumo_rl
sumolib==1.24.0
SuperSuit==3.10.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tinyscaler==1.2.8
torch==2.7.1
tqdm==4.67.1
traci==1.24.0
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
